+++ frontend/src/App.tsx
@@ -1,16 +1,10 @@
-import LoginAlertModal from '@/components/LoginAlertModal';
-import OAuthButton from '@/components/OAuthButton';
-import BottomSheet from '@/components/common/BottomSheet';
 import { Toast } from '@/components/common/Toast';
-import { OAUTH } from '@/constants';
-import { useAuth } from '@/hooks/useAuth';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { Outlet } from 'react-router-dom';
 import { useSpinnerStore } from '@/store/spinnerStore';
 import Spinner from '@/components/common/Spinner';
 // var console;
 function App() {
-    const { isLoggedIn } = useAuth();
     const { spinner } = useSpinnerStore();
     //TODO: 일시적인 배포시 console.log 제거 추가로 환경설정으로 빼줘야함ㄴ
     // if (process.env.NODE_ENV === 'production') {
@@ -34,35 +28,11 @@ function App() {
             return false;
         };
     }, []);
-    const [isLoginBottomSheetOpen, setLoginBottomSheetState] = useState(false);
-    const handleClose = () => {
-        setLoginBottomSheetState(false);
-    };
-    const handleToggle = (toggle: boolean) => {
-        setLoginBottomSheetState(toggle);
-    };
+
     return (
         <div className="flex flex-col w-full">
-            <div>
-                <Outlet />
-                <div>
-                    {!isLoggedIn && !isLoginBottomSheetOpen && (
-                        <LoginAlertModal isOpen={isLoginBottomSheetOpen} setToggle={handleToggle} />
-                    )}
-
-                    <BottomSheet isOpen={isLoginBottomSheetOpen} onClose={handleClose}>
-                        <BottomSheet.Body className="h-[230px]">
-                            <div className="flex flex-col items-center gap-3 mx-2 mt-4 mb-5">
-                                {OAUTH.map((oauth, index) => (
-                                    <OAuthButton key={index} provider={oauth.PROVIDER} name={oauth.NAME} />
-                                ))}
-                            </div>
-                        </BottomSheet.Body>
-                    </BottomSheet>
-                </div>
-            </div>
+            <Outlet />
             {spinner > 0 && <Spinner className="absolute bg-neutral-800/40 z-40" />}
-
             <Toast />
         </div>
     );
